+++ crates/libs/core/src/client/query_types.rs
@@ -6,14 +6,12 @@
 use crate::{
     iter::{FabricIter, FabricListAccessor},
     strings::HSTRINGWrap,
+    types::{HealthState, ServicePartitionInformation},
 };
 use bitflags::bitflags;
 use mssf_com::{
     FabricClient::{IFabricGetNodeListResult2, IFabricGetPartitionListResult2},
     FabricTypes::{
-        FABRIC_HEALTH_STATE, FABRIC_HEALTH_STATE_ERROR, FABRIC_HEALTH_STATE_INVALID,
-        FABRIC_HEALTH_STATE_OK, FABRIC_HEALTH_STATE_UNKNOWN, FABRIC_HEALTH_STATE_WARNING,
-        FABRIC_INT64_RANGE_PARTITION_INFORMATION, FABRIC_NAMED_PARTITION_INFORMATION,
         FABRIC_NODE_QUERY_RESULT_ITEM, FABRIC_NODE_QUERY_RESULT_ITEM_EX1,
         FABRIC_NODE_QUERY_RESULT_ITEM_EX2, FABRIC_PAGING_STATUS,
         FABRIC_QUERY_NODE_STATUS_FILTER_ALL, FABRIC_QUERY_NODE_STATUS_FILTER_DEFAULT,
@@ -27,11 +25,8 @@ use mssf_com::{
         FABRIC_QUERY_SERVICE_PARTITION_STATUS_NOT_READY,
         FABRIC_QUERY_SERVICE_PARTITION_STATUS_READY,
         FABRIC_QUERY_SERVICE_PARTITION_STATUS_RECONFIGURING, FABRIC_SERVICE_KIND_STATEFUL,
-        FABRIC_SERVICE_KIND_STATELESS, FABRIC_SERVICE_PARTITION_INFORMATION,
-        FABRIC_SERVICE_PARTITION_KIND_INT64_RANGE, FABRIC_SERVICE_PARTITION_KIND_INVALID,
-        FABRIC_SERVICE_PARTITION_KIND_NAMED, FABRIC_SERVICE_PARTITION_KIND_SINGLETON,
-        FABRIC_SERVICE_PARTITION_QUERY_DESCRIPTION, FABRIC_SERVICE_PARTITION_QUERY_RESULT_ITEM,
-        FABRIC_SINGLETON_PARTITION_INFORMATION,
+        FABRIC_SERVICE_KIND_STATELESS, FABRIC_SERVICE_PARTITION_QUERY_DESCRIPTION,
+        FABRIC_SERVICE_PARTITION_QUERY_RESULT_ITEM,
         FABRIC_STATEFUL_SERVICE_PARTITION_QUERY_RESULT_ITEM,
         FABRIC_STATELESS_SERVICE_PARTITION_QUERY_RESULT_ITEM, FABRIC_URI,
     },
@@ -255,29 +250,6 @@ impl From<&FABRIC_SERVICE_PARTITION_QUERY_RESULT_ITEM> for ServicePartition {
     }
 }
 
-// FABRIC_HEALTH_STATE
-#[derive(Debug, PartialEq)]
-pub enum HealthState {
-    Invalid,
-    Ok,
-    Warning,
-    Error,
-    Unknown,
-}
-
-impl From<&FABRIC_HEALTH_STATE> for HealthState {
-    fn from(value: &FABRIC_HEALTH_STATE) -> Self {
-        match *value {
-            FABRIC_HEALTH_STATE_INVALID => Self::Invalid,
-            FABRIC_HEALTH_STATE_OK => Self::Ok,
-            FABRIC_HEALTH_STATE_WARNING => Self::Warning,
-            FABRIC_HEALTH_STATE_ERROR => Self::Error,
-            FABRIC_HEALTH_STATE_UNKNOWN => Self::Unknown,
-            _ => Self::Invalid,
-        }
-    }
-}
-
 #[derive(Debug, PartialEq)]
 pub enum ServicePartitionStatus {
     Invalid,
@@ -312,88 +284,6 @@ pub struct StatefulServicePartition {
     pub last_quorum_loss_duration_in_seconds: i64,
     // TODO: reserved fields
     //pub Reserved: *mut core::ffi::c_void,
-}
-
-pub enum ServicePartitionInformation {
-    Invalid,
-    Singleton(SingletonPartitionInfomation),
-    Int64Range(Int64PartitionInfomation),
-    Named(NamedPartitionInfomation),
-}
-
-pub struct SingletonPartitionInfomation {
-    pub id: GUID,
-}
-
-#[derive(Debug)]
-pub struct Int64PartitionInfomation {
-    pub id: ::windows_core::GUID,
-    pub low_key: i64,
-    pub high_key: i64,
-}
-
-#[derive(Debug)]
-pub struct NamedPartitionInfomation {
-    pub id: ::windows_core::GUID,
-    pub name: ::windows_core::HSTRING,
-}
-
-impl From<&FABRIC_SINGLETON_PARTITION_INFORMATION> for SingletonPartitionInfomation {
-    fn from(value: &FABRIC_SINGLETON_PARTITION_INFORMATION) -> Self {
-        Self { id: value.Id }
-    }
-}
-
-impl From<&FABRIC_INT64_RANGE_PARTITION_INFORMATION> for Int64PartitionInfomation {
-    fn from(value: &FABRIC_INT64_RANGE_PARTITION_INFORMATION) -> Self {
-        Self {
-            high_key: value.HighKey,
-            id: value.Id,
-            low_key: value.LowKey,
-        }
-    }
-}
-
-impl From<&FABRIC_NAMED_PARTITION_INFORMATION> for NamedPartitionInfomation {
-    fn from(value: &FABRIC_NAMED_PARTITION_INFORMATION) -> Self {
-        Self {
-            id: value.Id,
-            name: HSTRINGWrap::from(value.Name).into(),
-        }
-    }
-}
-
-impl From<&FABRIC_SERVICE_PARTITION_INFORMATION> for ServicePartitionInformation {
-    fn from(value: &FABRIC_SERVICE_PARTITION_INFORMATION) -> Self {
-        match value.Kind {
-            FABRIC_SERVICE_PARTITION_KIND_SINGLETON => {
-                let raw = unsafe {
-                    (value.Value as *const FABRIC_SINGLETON_PARTITION_INFORMATION)
-                        .as_ref()
-                        .unwrap()
-                };
-                Self::Singleton(raw.into())
-            }
-            FABRIC_SERVICE_PARTITION_KIND_INT64_RANGE => {
-                let raw = unsafe {
-                    (value.Value as *const FABRIC_INT64_RANGE_PARTITION_INFORMATION)
-                        .as_ref()
-                        .unwrap()
-                };
-                Self::Int64Range(raw.into())
-            }
-            FABRIC_SERVICE_PARTITION_KIND_NAMED => {
-                let raw = unsafe {
-                    (value.Value as *const FABRIC_NAMED_PARTITION_INFORMATION)
-                        .as_ref()
-                        .unwrap()
-                };
-                Self::Named(raw.into())
-            }
-            FABRIC_SERVICE_PARTITION_KIND_INVALID => Self::Invalid,
-            _ => Self::Invalid,
-        }
-    }
 }
 
 impl From<&FABRIC_STATEFUL_SERVICE_PARTITION_QUERY_RESULT_ITEM> for StatefulServicePartition {
