+++ crates/libs/core/src/types/mod.rs
@@ -1,0 +1,4 @@
+// consists of type conversion of com api with safe rust wrappers.
+
+mod common;
+pub use common::*;
